# openapi.yaml
openapi: 3.0.3
info:
  title: Non-Linear Video Editor API
  description: |
    API for managing video editing projects, generating AI content (videos, images, audio),
    and handling user subscriptions.

    ## Authentication
    All endpoints require session-based authentication via Supabase Auth cookies.

    ## Rate Limiting
    The API uses a tiered rate limiting system:
    - Tier 1: 5 req/min (auth, payments)
    - Tier 2: 10 req/min (resource creation)
    - Tier 3: 30 req/min (read operations)
    - Tier 4: 60 req/min (general)
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
      description: Supabase authentication session cookie
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        field:
          type: string
          description: Field that caused the error
    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            limit:
              type: integer
              description: Rate limit maximum
            remaining:
              type: integer
              description: Remaining requests
            resetAt:
              type: integer
              format: int64
              description: Timestamp when limit resets
...
    Asset:
      type: object
      required:
        - id
        - project_id
        - user_id
        - type
        - storage_url
        - created_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [image, video, audio]
        source:
          type: string
          enum: [upload, genai]
        storage_url:
          type: string
          description: Internal storage URL
        mime_type:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
...
    VideoGenerationRequest:
      type: object
      required:
        - prompt
        - projectId
        - model
      properties:
        prompt:
          type: string
          minLength: 3
          maxLength: 1000
          description: Text description of the video
        projectId:
          type: string
          format: uuid
        model:
          type: string
          enum:
            - veo-3.1-generate-preview
            - veo-3.1-fast-generate-preview
            - veo-2.0-generate-001
            - seedance-1.0-pro
            - minimax-hailuo-02-pro
        aspectRatio:
          type: string
          enum: ['16:9', '9:16', '1:1', '4:3', '3:4']
        duration:
          type: number
          minimum: 1
          maximum: 10
        resolution:
          type: string
          enum: ['480p', '720p', '1080p']
        negativePrompt:
          type: string
          maxLength: 1000
        personGeneration:
          type: string
          enum: [dont_allow, allow_adult, allow_all]
        enhancePrompt:
          type: boolean
        generateAudio:
          type: boolean
        seed:
          type: integer
          minimum: 0
          maximum: 2147483647
        sampleCount:
          type: integer
          minimum: 1
          maximum: 4
        compressionQuality:
          type: integer
          minimum: 0
          maximum: 100
        imageAssetId:
          type: string
          format: uuid
    VideoGenerationResponse:
      type: object
      required:
        - operationName
        - status
        - message
      properties:
        operationName:
          type: string
          description: Operation ID for status polling
        status:
          type: string
          enum: [processing]
        message:
          type: string
...
    ImageGenerationRequest:
      type: object
      required:
        - prompt
        - projectId
      properties:
        prompt:
          type: string
          minLength: 3
          maxLength: 1000
        projectId:
          type: string
          format: uuid
        model:
          type: string
          default: imagen-3.0-generate-001
        aspectRatio:
          type: string
          enum: ['1:1', '3:4', '4:3', '9:16', '16:9']
        negativePrompt:
          type: string
          maxLength: 1000
        sampleCount:
          type: integer
          minimum: 1
          maximum: 8
        seed:
          type: integer
          minimum: 0
          maximum: 2147483647
        safetyFilterLevel:
          type: string
          enum: [block_most, block_some, block_few]
        personGeneration:
          type: string
          enum: [dont_allow, allow_adult, allow_all]
        addWatermark:
          type: boolean
        language:
          type: string
        outputMimeType:
          type: string
          enum: [image/png, image/jpeg]
    TTSRequest:
      type: object
      required:
        - text
        - projectId
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
        projectId:
          type: string
          format: uuid
        voiceId:
          type: string
          default: EXAVITQu4vr4xnSDxMaL
        modelId:
          type: string
          default: eleven_multilingual_v2
        stability:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        similarity:
          type: number
          minimum: 0
          maximum: 1
          default: 0.75
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - cookieAuth: []
paths:
...
  /video/generate:
    post:
      tags:
        - Video
      summary: Generate video
      description: Generate a video from text or image using AI
      operationId: generateVideo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoGenerationRequest'
      responses:
        '200':
          description: Video generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoGenerationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /image/generate:
    post:
      tags:
        - Image
      summary: Generate image
      description: Generate images from text using Google Imagen 3
      operationId: generateImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageGenerationRequest'
      responses:
        '200':
          description: Images generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /audio/elevenlabs/generate:
    post:
      tags:
        - Audio
      summary: Generate speech
      description: Generate speech from text using ElevenLabs TTS
      operationId: generateSpeech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
      responses:
        '200':
          description: Audio generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  asset:
                    $ref: '#/components/schemas/Asset'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '504':
          description: Request timeout (60s)
...